requirements.txt
amqp==5.1.1
asgiref==3.7.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-appconf==1.0.5
django-cors-headers==4.0.0
django-extensions==3.2.1
django-filter==23.2
django-imagekit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-yasg==1.21.5
eventlet==0.33.3
exceptiongroup==1.1.1
greenlet==2.0.2
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
packaging==23.1
pilkit==2.0
Pillow==9.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
PyJWT==2.7.0
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
redis==4.5.5
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
